refactor(sitelaunchpad): shift toast from dashboard -> launchpad
